course builder completed

+++ SERVER/controllers/Section.js
@@ -1,4 +1,5 @@
 const Section = require("../models/Section");
+const SubSection = require("../models/SubSection")
 
 const Course = require("../models/Course");
 const { findByIdAndUpdate } = require("../models/User");
@@ -60,7 +61,7 @@ exports.createSection = async (req, res) => {
 exports.updateSection = async (req, res) => {
   try {
     //data fecth
-    const { sectionName, sectionId } = req.body;
+    const { sectionName, sectionId ,courseId } = req.body;
 
     //validation
     if (!sectionName || !sectionId) {
@@ -76,11 +77,20 @@ exports.updateSection = async (req, res) => {
       { new: true }
     );
 
+      const course = await Course.findById(courseId)
+        .populate({
+          path: "courseContent",
+          populate: {
+            path: "subSection",
+          },
+        })
+        .exec();
+
     //return res
     return res.status(200).json({
       success: true,
-      message: "section updated successfully",
-      Data: section,
+      message: `section updated successfully ${section} `,
+      data: course,
     });
   } catch (error) {
     return res.status(500).json({
@@ -94,29 +104,40 @@ exports.updateSection = async (req, res) => {
 
 exports.deleteSection = async (req, res) => {
   try {
-    //get id
-    const { sectionId, courseId } = req.body;
+       const { sectionId, courseId } = req.body;
+       await Course.findByIdAndUpdate(courseId, {
+         $pull: {
+           courseContent: sectionId,
+         },
+       });
+       const section = await Section.findById(sectionId);
+       console.log(sectionId, courseId);
+       if (!section) {
+         return res.status(404).json({
+           success: false,
+           message: "Section not found",
+         });
+       }
+       // Delete the associated subsections
+       await SubSection.deleteMany({ _id: { $in: section.subSection } });
 
-    //verify kar lena isko
-    //TODO: Do we need to delete entry from Course Schema
-    //  const updatedCourse = await Course.findByIdAndUpdate(
-    //    { _id: courseId },
-    //    {
-    //      $pull: {
-    //        courseContent: courseId,
-    //      },
-    //    }
-    //  );
+       await Section.findByIdAndDelete(sectionId);
 
-    //delete
-    const response = await Section.findByIdAndDelete(sectionId);
+       // find the updated course and return it
+       const course = await Course.findById(courseId)
+         .populate({
+           path: "courseContent",
+           populate: {
+             path: "subSection",
+           },
+         })
+         .exec();
 
-    //return res
-    return res.status(200).json({
-      success: true,
-      message: "section deleted successfully",
-      //  Data: updatedCourse,
-    });
+       res.status(200).json({
+         success: true,
+         message: "Section deleted",
+         data: course,
+       });
   } catch (error) {
     return res.status(500).json({
       success: false,
